moved to using apollo react hooks
