+++ packages/newspringchurchapp/package.json
@@ -101,6 +101,7 @@
     "node-fetch": "2.6.0",
     "react": "16.8.3",
     "react-apollo": "2.5.5",
+    "react-apollo-hooks": "0.4.5",
     "react-native": "0.59.8",
     "react-native-actionsheet": "^2.4.2",
     "react-native-airplay-btn": "^1.0.6",
