+++ packages/newspringchurchapp/src/client/index.js
@@ -1,6 +1,7 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import { ApolloProvider } from 'react-apollo';
+import { ApolloProvider as ApolloHooksProvider } from 'react-apollo-hooks';
 import { ApolloClient } from 'apollo-client';
 import { ApolloLink } from 'apollo-link';
 import SplashScreen from 'react-native-splash-screen';
@@ -49,7 +50,11 @@ class ClientProvider extends PureComponent {
   }
 
   render() {
-    return <ApolloProvider {...this.props} client={client} />;
+    return (
+      <ApolloProvider {...this.props} client={client}>
+        <ApolloHooksProvider {...this.props} client={client} />
+      </ApolloProvider>
+    );
   }
 }
